:recycle: refactor: refactor ActionBar to a configurable stage
