+++ src/app/chat/features/ChatInput/ActionBar/ModelSwitch.tsx
@@ -1,0 +1,42 @@
+import { ActionIcon } from '@lobehub/ui';
+import { Dropdown } from 'antd';
+import { BrainCog } from 'lucide-react';
+import { memo } from 'react';
+import { useTranslation } from 'react-i18next';
+
+import { settingsSelectors, useGlobalStore } from '@/store/global';
+import { useSessionStore } from '@/store/session';
+import { agentSelectors } from '@/store/session/selectors';
+import { LanguageModel } from '@/types/llm';
+
+const ModelSwitch = memo(() => {
+  const { t } = useTranslation('setting');
+
+  const [model, updateAgentConfig] = useSessionStore((s) => {
+    const config = agentSelectors.currentAgentConfig(s);
+    return [config.model, s.updateAgentConfig];
+  });
+
+  const modelList = useGlobalStore(settingsSelectors.modelList);
+
+  return (
+    <Dropdown
+      menu={{
+        activeKey: model,
+        items: modelList.map((i) => ({ key: i, label: i })),
+        onClick: (e) => {
+          updateAgentConfig({ model: e.key as LanguageModel });
+        },
+        style: {
+          maxHeight: 400,
+          overflow: 'scroll',
+        },
+      }}
+      trigger={['click']}
+    >
+      <ActionIcon icon={BrainCog} placement={'bottom'} title={t('settingModel.model.title')} />
+    </Dropdown>
+  );
+});
+
+export default ModelSwitch;
